% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-url.R
\name{build_url}
\alias{build_url}
\title{Build URL.}
\usage{
build_url(to, from, message)
}
\arguments{
\item{to}{Number to send SMS to.}

\item{from}{Number to send SMS from.}

\item{message}{SMS message.}
}
\value{
URL character vector.
}
\description{
Build URL.
}
