% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalsend-token.R
\name{totalsend_token}
\alias{totalsend_token}
\title{Set or query Totalsend API token.}
\usage{
totalsend_token(token = NULL)
}
\arguments{
\item{token}{Totalsend API token.}
}
\value{
Token.
}
\description{
Set or query Totalsend API token.
}
\examples{
totalsend_token("jqbadf23aka6wsiv38fzyzi0ihu15zcfkkgl0rmnr1b4j7pg")
totalsend_token()
}
