% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalsend-username.R
\name{totalsend_username}
\alias{totalsend_username}
\title{Set or query Totalsend API username.}
\usage{
totalsend_username(username = NULL)
}
\arguments{
\item{username}{Totalsend username.}
}
\value{
Username.
}
\description{
Set or query Totalsend API username.
}
\examples{
totalsend_username("datawookie")
totalsend_username()
}
